✨ (processing): Rename CombineDataComponent to MergeDataComponent for better clarity and consistency 📝 (processing): Add new MergeDataComponent to combine data using different operations like Concatenate, Append, Merge, and Join
